backend/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Auth & JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validation & Config
pydantic==2.7.3
pydantic-settings==2.2.1

# File Handling
aiofiles==23.2.1
pillow==10.3.0

# PDF
reportlab==4.2.2
pypdf==4.2.0

# Scheduling
apscheduler==3.10.4

# Rate Limiting
slowapi==0.1.9

# HTTP Client (for file downloads, S3, etc.)
httpx==0.27.0
boto3==1.34.91  # AWS S3 support (optional)

# Logging
python-json-logger==2.0.7

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==4.1.0
httpx==0.27.0

# Development
black==24.3.0
flake8==7.0.0
isort==5.13.2
